+++ gdpr/src/main/java/com/kairos/persistence/repository/master_data/asset_management/hosting_provider/HostingProviderMDRepository.java
@@ -14,7 +14,7 @@ import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository
-public interface HostingProviderMDRepository extends JpaRepository<HostingProviderMD, Integer> {
+public interface HostingProviderMDRepository extends JpaRepository<HostingProviderMD, Long> {
 
     @Query(value = "SELECT hp FROM HostingProviderMD hp WHERE hp.countryId = ?1 and hp.deleted = ?2 and lower(hp.name) IN ?3")
     List<HostingProviderMD> findByCountryIdAndDeletedAndNameIn(Long countryId, boolean deleted, List<String> userNames);
@@ -22,16 +22,16 @@ public interface HostingProviderMDRepository extends JpaRepository<HostingProvid
     @Transactional
     @Modifying
     @Query(value = "update HostingProviderMD set name = ?1 where id= ?2")
-    Integer updateHostingProviderName(String name, Integer id);
+    Integer updateHostingProviderName(String name, Long id);
 
 
     @Transactional
     @Modifying
     @Query(value = "update HostingProviderMD set deleted = true where id = ?1 and countryId = ?2")
-    Integer deleteByIdAndCountryId(Integer id, Long countryId);
+    Integer deleteByIdAndCountryId(Long id, Long countryId);
 
     @Query(value = "SELECT hp FROM HostingProviderMD hp WHERE hp.id = ?1 and hp.countryId = ?2 and hp.deleted = ?3")
-    HostingProviderMD findByIdAndCountryIdAndDeleted(Integer id, Long countryId, boolean deleted);
+    HostingProviderMD findByIdAndCountryIdAndDeleted(Long id, Long countryId, boolean deleted);
 
 
     @Query(value = "SELECT new com.kairos.response.dto.common.HostingProviderResponseDTO(hp.id, hp.name, hp.organizationId, hp.suggestedDataStatus, hp.suggestedDate )  FROM HostingProviderMD hp WHERE hp.countryId = ?1 and hp.deleted = false order by createdAt desc")
